Renamed merging to session_directory, which is more appropriate.
