+++ rust/mdb_shard/src/lib.rs
@@ -2,8 +2,8 @@ pub mod cas_structs;
 pub mod error;
 pub mod file_structs;
 pub mod intershard_reference_structs;
-pub mod merging;
 pub mod serialization_utils;
+pub mod session_directory;
 pub mod set_operations;
 pub mod shard_file;
 pub mod shard_file_manager;
