+++ rust/xetblob/src/xet_repo.rs
@@ -17,7 +17,7 @@ use gitxetcore::merkledb_shard_plumb::{
     sync_session_shards_to_remote,
 };
 use gitxetcore::summaries_plumb::*;
-use mdb_shard::merging::consolidate_shards_in_session_directory;
+use mdb_shard::session_directory::consolidate_shards_in_session_directory;
 use mdb_shard::shard_file::MDB_SHARD_TARGET_SIZE;
 use merkledb::MerkleMemDB;
 use pointer_file::PointerFile;
